Слепок бд

+++ Database/dbo/Stored Procedures/uspRemoveCalendar.sql
@@ -1,5 +1,5 @@
 ﻿CREATE PROCEDURE [dbo].[uspRemoveCalendar]
-	@CalendarId int
+	@calendarId int
 AS
 BEGIN
 	SET NOCOUNT ON;
@@ -7,27 +7,24 @@ BEGIN
 	BEGIN TRANSACTION Transact
 		BEGIN TRY
 
-	----Default calendar and its events won't be removed
-		delete from UsersCalendars
-		where UsersCalendars.CalendarId in (select @CalendarId)
+		IF @calendarId NOT IN (SELECT CalendarDefaultId FROM Users)
+			BEGIN
+				delete from UsersCalendars
+				where UsersCalendars.CalendarId = @calendarId
 
-		delete from Calendars 
-		where Calendars.id = @CalendarId 
-		and @CalendarId not in (select distinct CalendarDefaultId from Users)
+				delete from Notification
+				where EventId in (select Id from Events where CalendarId = @calendarId)
 
-		delete from NotificationSchedule 
-		where EventScheduleId in 
-		(select id from EventSchedule 
-		where EventId in (
-		select Id from Events where CalendarId = @CalendarId))
+				--delete from EventInfinity where Eventid in (select id from Events where CalendarId = @CalendarId)
+				--delete from EventSchedule where Eventid in (select id from Events where CalendarId = @calendarId)
 
-		delete from NotificationInfinity
-		where EventId in (select Id from Events where CalendarId = @CalendarId)
+				delete from Events where CalendarId = @calendarId
 
-		delete from EventInfinity where Eventid in (select id from Events where CalendarId = @CalendarId)
-		delete from EventSchedule where Eventid in (select id from Events where CalendarId = @CalendarId)
+				delete from Calendars 
+				where Calendars.Id = @calendarId		
+				--Default calendar and its events won't be removed
+			END
 
-		delete from Events where CalendarId = @CalendarId
 	  COMMIT TRANSACTION Transact
 	  END TRY
 	  BEGIN CATCH
